FranklinContent.tag_link — Function

The HTML link to a page display all the tags.

FranklinContent.hfun_addtags — Function

Tag list to display in the post footer.

FranklinContent.hfun_insert_img — Function

IMG html tag, where the alt attribute defaults to the image name.

FranklinContent.hfun_about_place — Function

The evaluated place.html file located in the :layout franklin folder.

FranklinContent.hfun_recent_posts — Function

HTML list of recent posts. arg1: the root path of all the posts arg2: how many posts to display

FranklinContent.hfun_star — Function

A colored (var(–color)) star.

FranklinContent.hfun_tags_cloud — Function

Tag cloud, tags with font size dependent on the number of posts that use it.

FranklinContent.hfun_insert_path — Function

Inserts a file relative to the current page.

FranklinContent.iter_posts — Function

All the posts files (.md) in the :posts_path directory.

FranklinContent.page_content — Function

Process a franklin markdown page and return the output HTML.

FranklinContent.post_crumbs — Function

Breadcrumbs for the current page being evaluated.

FranklinContent.hfun_insert_bio — Function

The (modified) biographic details from the field bio.md located in the :assets franklin folder.

FranklinContent.hfun_post_title — Function

HTML for the page title, using rss_description for the subtitle.

FranklinContent.post_link — Function

The absolute or relative post link from a file path.

FranklinContent.hfun_tag_title — Function

The tag of the current page, none otherwise.

FranklinContent.hfun_addcomments — Function

The utteranc.es comments widget (if the page is a post).

FranklinContent.hfun_canonical_link — Function

The html link tag for the current canonical url.

FranklinContent.lx_fun — Function

Wrapper to call an arbitrary function as an latex function.

FranklinContent.hfun_taglist_desc — Function

All the pages for a particular tag.

FranklinContent.hfun_color — Function

Text wrapped in an HTML tag with a specified color.

FranklinContent.is_index — Function

Check if page is an index page.

FranklinContent.is_post — Function

Check if page is a post.

FranklinContent.is_tag — Function

Check if page is a tags page.

FranklinContent.hfun_editedpage — Function

Add edited date at appropriate pages.

FranklinContent.hfun_icon_tag — Function

Wraps the icon_tags dict.